Initialize local variable to placate gcc
The problem is that there are other codepage groups beside 12000 and 12001.
Now, i haven't checked the code thoroughly, but it appears that winiconv
*first* checks the codepage, then assigns one of the conversion functions
(such as utf32_mbtowc) to a placeholder variable, which later is called.
So it seems that it's guaranteed that utf32_mbtowc() will only be called
for codepage groups 12000 and 12001 (but gcc doesn't get it, sadly).

Thus initializing uint wc to 0xD800 just placates gcc, wc never actually
gets used without being set.